update real time claim

+++ src/database/models/user.js
@@ -40,7 +40,7 @@ module.exports = (sequelize, DataTypes) => {
       first_name: { type: DataTypes.STRING, allowNull: true },
       last_name: { type: DataTypes.STRING, allowNull: true },
       username: { type: DataTypes.STRING, allowNull: true },
-      coin_balance: { type: DataTypes.INTEGER, allowNull: true },
+      coin_balance: { type: DataTypes.FLOAT, allowNull: true },
       level_point: { type: DataTypes.INTEGER, allowNull: true },
       energy_point: { type: DataTypes.INTEGER, allowNull: true },
       energy_size_level: { type: DataTypes.INTEGER, allowNull: true },
